- Use relative sizes for buttons - highlight button whilst slowed - Only allow slow if not already slowed - use OrthographicCamera - resize font based on screen size - resize orbs based on screen size

+++ core/src/main/java/net/scriptgate/orbital/tap/Button.java
@@ -4,7 +4,7 @@ public class Button extends Rectangle{
 
     public boolean hover = false;
 
-    public Button(int x, int y, int height, int width) {
+    public Button(float x, float y, float height, float width) {
         super(x, y, height, width);
     }
 }
